service: on network disconnect, remove profile from supplicant.
To avoid MAC address leak, on network disconnection, MAC is randomized.
This leads to VDEV reset in driver and thus driver losses associated
PMKSA/PMKID details for network profile. Though framework sets pmksa
entry to supplicant, supplicant doesn't set it driver as it had already
set in last connection. Thus, on reconnection, full authentication
is performed again.

This commit is to remove the network on disconnection to allow
re-plumbing of PMKSA cache to driver.

Change-Id: Ifa3ed770ea28738cdc6f0d723d2cb8349d078cd7
CRs-Fixed: 3001200