Make InterfaceDestroyedListener run on the same thread as Manager which create this interface
When create the Iface, the null handler makes the
InterfaceDestroyedListener running on the other Thread when creating
new Iface, as new interface manager may running on a different thread
Which leads to a deadlock.

Make sure a valid handler must be set when create an Iface with
InterfaceDestroyedListener.

Test: atest FrameworksWifiTests
Bug: 199792691
Change-Id: I5bb64bd1fcb1cf88a2fa4d924052fea6e9e9a331
CRs-Fixed: 3035002
(Cherry-picked from 161347e9e28881ae6f55b0397b500cdcedd0c45c)